requirements.txt
discord.py
pynacl
yt_dlp
python-dotenv

